+++ src/Runtime/FixtureCollectionBuilder.php
@@ -2,24 +2,19 @@
 
 namespace AKlump\FixtureFramework\Runtime;
 
-use AKlump\FixtureFramework\Helper\FixtureInstantiator;
-
 class FixtureCollectionBuilder {
 
   public function __construct(
-    private array|RunOptions $options,
-    private RunContextValidator $validator
+    private readonly FixtureInstantiator $instantiator,
   ) {
   }
 
   public function __invoke(array $definitions) {
     $store = new RunContextStore();
 
-    return array_map(fn(array $definition) => (new FixtureInstantiator())(
+    return array_map(fn(array $definition) => ($this->instantiator)(
       $definition,
-      $this->options,
       $store,
-      $this->validator
     ), $definitions);
   }
 }